refine tests

+++ test/Fluent/FluentHttpLoggerTest.php
@@ -7,33 +7,49 @@ class FluentHttpLoggerTest extends PHPUnit_Framework_TestCase
 
     function setUp()
     {
-        $this->pidFilePath = tempnam('/tmp', 'fluent.pid');
-        $command = sprintf('bash --login -c "fluentd -c %s -d %s"', dirname(__FILE__) . '/fixture/fluent.conf', $this->pidFilePath);
+        $this->pidFilePath = tempnam('/tmp', 'fluentd.pid');
+        $command = sprintf('bash --login -c "fluentd -c %s -d %s"', dirname(__FILE__) . '/fixture/fluent.http.conf', $this->pidFilePath);
         exec($command);
-        sleep(1);
     }
 
     function tearDown()
     {
-        posix_kill(file_get_contents($this->pidFilePath), 9);
         if (file_exists($this->pidFilePath)) {
+            $command = sprintf('kill %s', file_get_contents($this->pidFilePath));
+            exec($command);
             unlink($this->pidFilePath);
         }
+
+        $this->cleanup();
     }
 
     function testSend()
     {
-        $logger = new FluentHttpLogger('myapp', 'localhost');
-        $logger->send(array('user' => 'abc', 'password' => 'def'));
+        $this->cleanup();
 
-        $logPaths = glob('/tmp/fluent.phpunit.*');
-        foreach ($logPaths as $logPath) {
-            $contents = file_get_contents($logPath);
-            unlink($logPath);
-            $this->assertSame(32, strpos($contents, '{"user":"abc","password":"def"}'));
-            return;
+        $logger = new FluentHttpLogger('phpunit.fluent', 'localhost');
+        $logger->send(array('user' => 'abc', 'password' => 'def'));
+        sleep(1);
+
+        $logPaths = glob('/tmp/fluentd.phpunit.*');
+        if (count($logPaths) !== 1) {
+            $this->fail();
         }
-        unlink($logPath);
-        $this->fail();
+        $logPath = $logPaths[0];
+        $contents = file_get_contents($logPath);
+        if (preg_match('/^(?P<time>.*)\t(?P<tag>.*)\t(?P<data>.*)$/', $contents, $matches) !== 1) {
+            $this->fail();
+        }
+        $this->assertSame('{"user":"abc","password":"def"}', $matches['data']);
+    }
+
+    private function cleanup()
+    {
+        $logPaths = glob('/tmp/fluentd.phpunit.*');
+        if (!empty($logPaths)) {
+            foreach ($logPaths as $logPath) {
+                unlink($logPath);
+            }
+        }
     }
 }
